Add files via upload
This GitHub repository encapsulates a comprehensive guide to text classification using BERT in NLP. With a focus on spam and non-spam email classification, users are equipped with a thorough understanding of data preprocessing, BERT embeddings, model building, training, evaluation, and inference. The tutorial serves as a valuable resource for NLP enthusiasts, data scientists, and machine learning practitioners, offering not just code snippets but a rich learning experience.